# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于 Go + GORM 的博客系统后端",
        "title": "我的博客 API",
        "contact": {},
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/app-info": {
            "get": {
                "description": "获取当前后端应用的版本、环境、运行状态等基础信息",
                "tags": [
                    "基础模块"
                ],
                "summary": "获取系统信息",
                "responses": {
                    "200": {
                        "description": "返回应用配置信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.JsonResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.AppInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/article": {
            "get": {
                "tags": [
                    "文章模块"
                ],
                "summary": "获取文章列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.JsonResult"
                        }
                    }
                }
            },
            "put": {
                "description": "根据文章对象中的 ID 更新标题、内容、分类等信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章模块"
                ],
                "summary": "更新文章",
                "parameters": [
                    {
                        "description": "需要更新的文章对象",
                        "name": "article",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Article"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功后的文章详情",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.JsonResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ArticleVO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "提交文章标题、内容等信息创建新篇章",
                "tags": [
                    "文章模块"
                ],
                "summary": "新增文章",
                "parameters": [
                    {
                        "description": "文章内容对象",
                        "name": "article",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Article"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回创建成功的文章详情",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.JsonResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ArticleVO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "根据文章 ID 永久删除文章",
                "tags": [
                    "文章模块"
                ],
                "summary": "删除文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章 ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回被删除的文章 ID",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.JsonResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/article/{id}": {
            "get": {
                "description": "根据文章 ID 查询文章的详细内容及分类信息",
                "tags": [
                    "文章模块"
                ],
                "summary": "获取文章详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.JsonResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ArticleVO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "datatypes.JSONType-model_AppContent": {
            "type": "object"
        },
        "model.AppInfo": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "使用 JSONType 绑定上面的结构体",
                    "allOf": [
                        {
                            "$ref": "#/definitions/datatypes.JSONType-model_AppContent"
                        }
                    ]
                },
                "id": {
                    "type": "integer"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "model.Article": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "description": "必须大写开头！后面反引号里的才是数据库真实的列名",
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "typeId": {
                    "type": "string"
                }
            }
        },
        "model.ArticleVO": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "description": "必须大写开头！后面反引号里的才是数据库真实的列名",
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "typeId": {
                    "type": "string"
                },
                "typeName": {
                    "description": "扩展字段",
                    "type": "string"
                }
            }
        },
        "utils.JsonResult": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务约定：0 为成功，-1 为失败",
                    "type": "integer"
                },
                "data": {
                    "description": "业务数据"
                },
                "msg": {
                    "description": "错误描述或提示信息",
                    "type": "string"
                }
            }
        }
    }
}